refactor: rename module_default by default (model)
