+++ src/core/__templates__/models/default.py
