+++ src/core/abi/agents/EntitytoSPARQLAgent.py
@@ -136,7 +136,7 @@ def create_agent(
     # Define model based on AI_MODE
     ai_mode = secret.get("AI_MODE")  # Default to cloud if not set
     if ai_mode == "airgap":
-        from src.core.abi.models.module_default import airgap_model as model
+        from src.core.abi.models.default import airgap_model as model
     else:
         from src.core.chatgpt.models.o3_mini import model
