+++ src/marketplace/applications/powerpoint/models/default.py
